Implemented "long press to quiver"
The next step to getting tour deletion to work was to make the collection view behave like Springboard, which lets you long press on apps/folders to bring up the ability to delete items.  Long pressing on a cell now puts the collection view into deletion mode, showing the delete button on the cells as well as causing the cells to quiver.  Tapping anywhere in the collection view (except for the cells) disables deletion mode.

Next stop: handling tapping the delete button.  It will bring up an alert controller asking the user if they're sure they want to delete the selected tour.