Implemented deleting tours
Tours can now be deleted from the dashboard.  The user would first long press on any of the tours, tap on the delete button of the tour they want to delete, at which time they are presented with a confirmation.  If they select Delete, the tour will be removed from Core Data and the cell will be removed from the collection view.