Implemented the Dashboard (implements #11) (#85)
* Adding Dashboard changes

* Removed DZNEmptyDataSet pod

* Added some add tour functionality

Add Tour is not yet fully functional, but I have added the ability to add a tour, which is then saved into the persistent store.

* Trying to get add tour cell to work

* Trying something

* Got the dashboard working

It turns out that my approach to getting the Add Tour cell tacked onto the end of the collection view has been right all along...  I just forgot to set the class for my Add Tour cell to AddTourCollectionViewCell in the storyboard...

* Added cell delete button; cell shaking

Started adding code to make the dashboard function similar to Springboard.  Added a delete button to the cells as well as animation code to make the cells shake.

* Implemented "long press to quiver"

The next step to getting tour deletion to work was to make the collection view behave like Springboard, which lets you long press on apps/folders to bring up the ability to delete items.  Long pressing on a cell now puts the collection view into deletion mode, showing the delete button on the cells as well as causing the cells to quiver.  Tapping anywhere in the collection view (except for the cells) disables deletion mode.

Next stop: handling tapping the delete button.  It will bring up an alert controller asking the user if they're sure they want to delete the selected tour.

* Implemented deleting tours

Tours can now be deleted from the dashboard.  The user would first long press on any of the tours, tap on the delete button of the tour they want to delete, at which time they are presented with a confirmation.  If they select Delete, the tour will be removed from Core Data and the cell will be removed from the collection view.

* Removed big titles

I have removed the big titles on the Dashboard and Add Tour views and embed both view controllers in navigation controllers.