Location List: Add locations to Tours (#87)
* Basic Location List and Map.

* Basic add locations to new tour.

* Fixed project problem

* Trying to pass Tour between VCs

* Tour object now being passed to LocationListVC

In order to correctly pass the new Tour object from the AddTourVC to LocationListVC, I needed to basically re-fetch the object from the store after it has been successfully saved.  To ensure that I fetch the correct object (currently, Tour objects could be given the same title), I added a new UUID property to the Tour object.  That UUID is generated when the user presses the add tour button and saved with the new tour object.  In the completion handler, if the save was successful, we perform the segue, which first fetches the new tour object so that it can be passed along.

* Fixing expliclity unwrapped optionals

* Minor logic change

Removed the tour property from the AddTourVC class because it's not necessary.