Fixed working with Tours in MagicalRecordModelService
I've started to go through the MagicalRecordModelService to get it working correctly.  First stop was working with Tours.  I've made the appropriate changes to create tours with an NSUUID, a new property added to the Tour entity.  I've also added functions to find tours by UUID as well as delete tours.

While I was in there, I modified the data services to not use a singleton.