Fixed MagicalRecordModelService (mostly); UI improvements (#92)
* Fixed working with Tours in MagicalRecordModelService

I've started to go through the MagicalRecordModelService to get it working correctly.  First stop was working with Tours.  I've made the appropriate changes to create tours with an NSUUID, a new property added to the Tour entity.  I've also added functions to find tours by UUID as well as delete tours.

While I was in there, I modified the data services to not use a singleton.

* More work on MagicalRecordService; changes to LocationList

I have been able to get most of the MagicalRecordModelService working.  There are still a couple of functions that need to be addressed.  I have also modified the layout of the LocationList view to give it a more native feel.  I removed the placeholder image that was there and am now using my dynamically loaded random placeholder image.  I need to go back and put the "Closest to Me" sort option back in; I haven't figured out how to display that yet.  I'm thinking it will be another bar button item.

I've added a custom font (Ubuntu) and have started to use that font throughout the app.  This is mainly for my appeasement as I don't really like the system font.  We can change it later if others don't like it.

On the CurrentTour view, I've implemented displaying the tours that have been saved for that tour both in the table view and map view.  I've also added an Add Location bar button which brings the user back to the LocationList view to add more locations.

* Fixed some code review defects

Replaced some unsafe force unwrapped optionals, refactored a function in the MagicalRecordModelService

* More defect fixes

Add calls to appropriate superclass methods where needed (prepareForSegue(), viewDidAppear())

* Without these changes, the app crashes

* Adding one last TODO