Fixed dashboard crash (fixes #88) (#96)
Dashboard was crashing when I deleted the last tour on the dashboard because of how I was adding a new "Add Tour" collection view.  Instead of manually removing the cell from the collection view, I am simply calling reloadData() now, which takes care of that crash.

On top of the crash, when implementing the above change, the "AddTourView" that is normally shown when there are no tours on the Dashboard wasn't being shown after that last tour was deleted.  I added code to numberOfItemsInSection() that will show the view when the number of tours is zero and the tour is currently hidden.