Created Location Detail View
This is an initial go at the Location Detail View.  There are still some things that need to be worked on.  The presentation of information is somewhat lacking and could do with some UX love.  We also need to be able to add a location to a tour from this view (if the user is coming from the LocationList view).  And, of course, the Image View at the top needs to be implemented to allow multiple location photos.