Fixed a memory issue due to improper view controller management (#104)
* Temporary commit

* Updated Cocoapods; fixed VC problems

There was an issue where tapping "Home" from the CurrentTour view controller was incorrectly creating new Dashboard view controllers, causing the app to use more and more memory.  To remedy this, I added an unwind segue to DashboardVC and now, tapping on "Home" triggers the unwind segue instead, to better manage the view transition.

During all this, I ended up updating to CocoaPods v1.2, which is why there are so many Pod-related changes in this commit.

* Refactored dashboard data source

Because I ended up duplicating code in both viewDidLoad() and viewWillAppear(), I refactored that code into a separate function.